+++ nisystemlink/clients/spec/utilities/__init__.py
@@ -1,3 +1,3 @@
-from ._dataframe_utility import get_specs_dataframe
+from ._dataframe_utilities import get_specs_dataframe
 
 # flake8: noqa
