+++ nisystemlink/clients/spec/utilities/_constants.py
@@ -1,0 +1,7 @@
+CONDITION_COLUMN_HEADER = "condition_"
+
+PROPERTY_COLUMN_HEADER = "properties."
+
+KEYWORDS_COLUMN_HEADER = "keywords"
+
+DEFAULT_QUERY_SPECS_TAKE = 1000
